+++ src/RhinoInside.Revit.GH/Components/Structure/AddAnalyticalMember.cs
@@ -14,7 +14,7 @@ namespace RhinoInside.Revit.GH.Components.Structure
 #endif
 
   [ComponentVersion(introduced: "1.27"), ComponentRevitAPIVersion(min: "2023.0")]
-  public class AddAnalyticalMember : BaseAnalyticalComponent
+  public class AddAnalyticalMember : AddAnalyticalElement
   {
     public override Guid ComponentGuid => new Guid("88AD5522-B3AD-4A67-AB96-3D90249BA215");
 #if REVIT_2023